refactor client to extract the request client into its own struct and create a key provider interface
