+++ client.go
@@ -1,31 +1,23 @@
 package goclash
 
 import (
-	"errors"
-	"fmt"
 	"net/http"
-	"slices"
 	"strconv"
-	"sync"
-	"time"
 
 	"github.com/bytedance/sonic"
 	"github.com/go-resty/resty/v2"
 )
 
-type Client struct {
-	accounts []*APIAccount
-	rc       *resty.Client
-	ipAddr   string
-	keyIndex APIKeyIndex
-	cache    *Cache
-	mu       sync.Mutex
+type KeyProvider interface {
+	GetKey() string
+	// Does all the necessary work to update the key provider, making the keys valid ones again
+	RevalidateKeys() error
 }
 
-var defaultHeaders = map[string]string{
-	"Accept":       "application/json",
-	"Content-Type": "application/json",
-	"User-Agent":   "goclash",
+type Client struct {
+	keys  KeyProvider
+	rc    *RequestsClient
+	cache *Cache
 }
 
 func newClient(creds Credentials) (*Client, error) {
@@ -39,16 +31,18 @@ func newClient(creds Credentials) (*Client, error) {
 		})
 	}
 
-	client := &Client{
+	rc := &RequestsClient{resty.New()}
+	keys := &AccountsKeyProvider{
 		accounts: accounts,
-		rc:       resty.New(),
-		cache:    newCache(),
+		rc:       rc,
+	}
+	client := &Client{
+		keys:  keys,
+		rc:    rc,
+		cache: newCache(),
 	}
 
-	if err := client.updateIPAddr(); err != nil {
-		return nil, err
-	}
-	if err := client.updateAccounts(); err != nil {
+	if err := keys.RevalidateKeys(); err != nil {
 		return nil, err
 	}
 
@@ -82,10 +76,7 @@ func (h *Client) do(method, url string, req *resty.Request, retry bool) ([]byte,
 		}
 
 		if clientErr.APIError.Reason == ReasonInvalidIP {
-			if err = h.updateIPAddr(); err != nil {
-				return nil, err
-			}
-			if err = h.updateAccounts(); err != nil {
+			if err = h.keys.RevalidateKeys(); err != nil {
 				return nil, err
 			}
 			return h.do(method, url, req, false)
@@ -95,185 +86,12 @@ func (h *Client) do(method, url string, req *resty.Request, retry bool) ([]byte,
 	return nil, clientErr
 }
 
-func (h *Client) updateIPAddr() error {
-	res, err := h.rc.R().Get(IPifyEndpoint)
-	if err != nil {
-		return err
-	}
-
-	body := string(res.Body())
-	if res.StatusCode() != http.StatusOK {
-		return errors.New(body)
-	}
-	if body == "" {
-		return errors.New("couldn't get IP address")
-	}
-	if body == h.ipAddr {
-		return nil
-	}
-
-	h.mu.Lock()
-	h.ipAddr = body
-	h.mu.Unlock()
-	return nil
-}
-
-func (h *Client) login(account *APIAccount) error {
-	res, err := h.newDefaultRequest().SetBody(account.Credentials).Post(DevLoginEndpoint.URL())
-	if err != nil {
-		return err
-	}
-
-	if res.StatusCode() != http.StatusOK {
-		return errors.New(string(res.Body()))
-	}
-
-	return sonic.Unmarshal(res.Body(), &account)
-}
-
-func (h *Client) updateAccounts() error {
-	for _, account := range h.accounts {
-		if err := h.login(account); err != nil {
-			return err
-		}
-		if err := h.updateAccountKeys(account); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-// getAccountKeys retrieves the API keys for the given account and sets APIAccount.Keys.
-func (h *Client) getAccountKeys(account *APIAccount) error {
-	res, err := h.newDefaultRequest().Post(DevKeyListEndpoint.URL())
-	if err != nil {
-		return err
-	}
-
-	if res.StatusCode() != http.StatusOK {
-		return errors.New(string(res.Body()))
-	}
-
-	var body *KeyListResponse
-	if err = sonic.Unmarshal(res.Body(), &body); err != nil {
-		return err
-	}
-
-	h.mu.Lock()
-	for i := range body.Keys {
-		account.Keys[i] = body.Keys[i]
-	}
-	h.mu.Unlock()
-	return nil
-}
-
-func (h *Client) updateAccountKeys(account *APIAccount) error {
-	if err := h.getAccountKeys(account); err != nil {
-		return err
-	}
-
-	errChan := make(chan error, keysPerAccount)
-	var freeKeyIndexes []int
-	var wg sync.WaitGroup
-	for i := 0; i < keysPerAccount; i++ {
-		if account.Keys[i] == nil {
-			freeKeyIndexes = append(freeKeyIndexes, i)
-			continue
-		}
-		if !slices.Contains(account.Keys[i].CidrRanges, h.ipAddr) {
-			wg.Add(1)
-			go func(key *APIKey, i int) {
-				defer wg.Done()
-				if err := h.revokeAccountKey(key); err != nil {
-					errChan <- err
-					return
-				}
-				if err := h.createAccountKey(account, i); err != nil {
-					errChan <- err
-					return
-				}
-			}(account.Keys[i], i)
-		}
-	}
-	wg.Wait()
-
-	for i := range freeKeyIndexes {
-		wg.Add(1)
-		go func(i int) {
-			defer wg.Done()
-			if err := h.createAccountKey(account, i); err != nil {
-				errChan <- err
-			}
-		}(i)
-	}
-	wg.Wait()
-
-	if len(errChan) > 0 {
-		return <-errChan
-	}
-	return nil
-}
-
-func (h *Client) createAccountKey(account *APIAccount, index int) error {
-	desc := fmt.Sprintf("Created at %s by goclash", time.Now().UTC().Round(time.Minute).String())
-	key := &APIKey{
-		Name:        "goclash",
-		Description: desc,
-		CidrRanges:  []string{h.ipAddr},
-		Scopes:      []string{"clash"},
-	}
-	res, err := h.newDefaultRequest().SetBody(key).Post(DevKeyCreateEndpoint.URL())
-	if err != nil {
-		return err
-	}
-
-	if res.StatusCode() != http.StatusOK {
-		return errors.New(string(res.Body()))
-	}
-
-	var keyRes *CreateKeyResponse
-	if err = sonic.Unmarshal(res.Body(), &keyRes); err != nil {
-		return err
-	}
-
-	h.mu.Lock()
-	defer h.mu.Unlock()
-	account.Keys[index] = keyRes.Key
-	return nil
-}
-
-func (h *Client) revokeAccountKey(key *APIKey) error {
-	payload := map[string]string{"id": key.ID}
-	res, err := h.newDefaultRequest().SetBody(payload).Post(DevKeyRevokeEndpoint.URL())
-	if err != nil {
-		return err
-	}
-
-	if res.StatusCode() != http.StatusOK {
-		return errors.New(string(res.Body()))
-	}
-	return nil
-}
-
-func (h *Client) getKey() string {
-	h.mu.Lock()
-	defer h.mu.Unlock()
-
-	key := h.accounts[h.keyIndex.AccountIndex].Keys[h.keyIndex.KeyIndex]
-	if h.keyIndex.KeyIndex == len(h.accounts[h.keyIndex.AccountIndex].Keys)-1 {
-		h.keyIndex.AccountIndex = (h.keyIndex.AccountIndex + 1) % len(h.accounts)
-	}
-	h.keyIndex.KeyIndex = (h.keyIndex.KeyIndex + 1) % len(h.accounts[h.keyIndex.AccountIndex].Keys)
-	return key.Key
-}
-
 func (h *Client) newDefaultRequest() *resty.Request {
-	return h.rc.R().SetHeaders(defaultHeaders)
+	return h.rc.NewDefaultRequest()
 }
 
 func (h *Client) withAuth(req *resty.Request) *resty.Request {
-	return req.SetAuthToken(h.getKey())
+	return req.SetAuthToken(h.keys.GetKey())
 }
 
 func (h *Client) withPaging(r *resty.Request, params *PagingParams) *resty.Request {
